refactor ( website-builder ): Refactor website-builder to typescript
The index.js file is now refactored to index.ts and a tsconfig.json has
been added along with required type packages as dev dependencies

closes #1054
