+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "description": "This repository is our website deploy and update tool to minimize github api queries.",
   "main": "index.js",
-  "type" : "module",
   "scripts": {
     "start": "node index.js",
     "lint": "eslint --ext=.ts --debug .",
@@ -36,9 +35,12 @@
     "express": "^4.19.2"
   },
   "devDependencies": {
-    "@idrinth-api-bench/eslint-config": "https://github.com/idrinth-api-bench/eslint-config#setup-base-config",
     "@commitlint/cli": "^19.3.0",
-    "simple-git-hooks": "^2.11.1"
+    "@idrinth-api-bench/eslint-config": "https://github.com/idrinth-api-bench/eslint-config#setup-base-config",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.12.12",
+    "simple-git-hooks": "^2.11.1",
+    "typescript": "^5.4.5"
   },
   "engineStrict": true,
   "engines": {
